@@ -5,9 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha",
-    "start": "nodemon index.js",
-    "lint": "eslint ./",
-    "lintfix": "eslint ./ --fix"
+    "start": "nodemon dist/index.js",
+    "lint": "eslint --ext .ts ."
   },
   "repository": {
     "type": "git",
@@ -29,15 +28,21 @@
     "multer": "^1.4.2"
   },
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
+    "@types/express": "^4.17.7",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/mongodb": "^3.5.26",
+    "@types/multer": "^1.4.4",
+    "@types/node": "^14.6.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "chai": "^4.2.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.7.0",
     "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-promise": "^4.2.1",
     "mocha": "^7.2.0",
     "nodemon": "^1.19.4",
-    "sinon": "^8.1.1"
+    "sinon": "^8.1.1",
+    "typescript": "^3.9.7"
   }
 }
