@@ -1,84 +1,90 @@
 /* eslint-disable no-useless-escape */
-const { validationResult } = require("express-validator");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const { getDb } = require("../util/database");
+import { Request, Response, NextFunction } from 'express';
+import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
 
-exports.signup = (req, res) => {
+import { validationResult } from 'express-validator'
+import { getDb } from '../util/database'
+
+import { wcError } from '../custom-types'
+
+export function signup (req: Request, res: Response): void {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     res.status(422).json({
-      message: "Validation failed",
-      errors: errors.array()
+      message: 'Validation failed',
+      errors: errors.array(),
     });
-    throw new Error("Validation failed Error");
+    throw new Error('Validation failed Error');
   }
-  const { userName, password, name, lastName, email } = req.body;
+  const {
+    userName, password, name, lastName, email,
+  } = req.body;
   bcrypt
     .hash(password, 12)
-    .then(passwordEnc => {
+    .then((passwordEnc) => {
       (async () => {
         const user = {
           userName,
           password: passwordEnc,
           name,
           lastName,
-          email
+          email,
         };
         const db = getDb().db();
         const insertOneWriteOpResultObject = await db
-          .collection("users")
+          .collection('users')
           .insertOne(user);
         const insertedResult = insertOneWriteOpResultObject.ops[0];
         res.status(200).json({
-          insertedResult
+          insertedResult,
         });
         return insertedResult;
       })();
       return true;
     })
-    .catch(error => {
+    .catch((error) => {
       // eslint-disable-next-line no-console
-      console.log("error", error);
+      console.log('error', error);
     });
-};
+}
 
-exports.login = async (req, res, next) => {
+export async function login (req: Request, res: Response, next: NextFunction): Promise<void> {
   try {
     const { email } = req.body;
     const { password } = req.body;
     const db = getDb().db();
-    const user = await db.collection("users").findOne({ email });
+    const user = await db.collection('users').findOne({ email });
     if (user === null) {
-      const customErrorMessage = "Email doesn't exist.";
-      const error = new Error(customErrorMessage);
-      error.statusCode = 400;
-      error.customErrorMessage = customErrorMessage;
+      const error:wcError = {
+        statusCode: 400,
+        customErrorMessage: "Email doesn't exist."
+      }
       throw error;
     }
     const loadedUser = user;
     const userPasswordAndStorePasswordAreIqual = await bcrypt.compare(
       password,
-      user.password
+      user.password,
     );
     if (!userPasswordAndStorePasswordAreIqual) {
-      const error = new Error("Password is wrong");
-      error.statusCode = 400;
+      const error:wcError = {
+        statusCode: 400,
+        customErrorMessage: "Password is wrong"
+      }
       throw error;
     }
     const token = jwt.sign(
       {
         // eslint-disable-next-line no-underscore-dangle
-        userId: loadedUser._id.toString()
+        userId: loadedUser._id.toString(),
       },
-      // eslint-disable-next-line prettier/prettier
       "P2bQM9L2r4b$}3<X)_cU_+F7-:}'Y-]e(gPH::?^*YW,x5<3*Zrfy=zZ^.K!aJ,6!S&UJ;fTCmnz}.>\,^mtK8d{jw8a88z`yARW@b78K8+TxNg6{Eg?wPXZ3%:fjZY3;V[dX#Y7t\K9]sXXXud+mFjqtM#[q\(UL#.c-L@M99wfJ2RsaSj\Q7x/Gwnmnk+c6-Jb.n-&:J'jzS:fGDa*e6aw;W(u$!R6~cbed}tWw7.[g]yTzx@56Hhz}2{L`6%;",
-      { expiresIn: "1hr" }
+      { expiresIn: '1hr' },
     );
     // eslint-disable-next-line no-underscore-dangle
     res.status(200).json({ token, userId: loadedUser._id.toString() });
   } catch (error) {
-    console.log("error 85", error);
     next(error);
   }
-};
+}
